fix(security): timing attack, stale sessions, tenant isolation, weight truncation
- AuthController: constant-time login — always run password_verify with
  dummy hash when user not found, preventing timing-based user enumeration
- AuthMiddleware: periodic session re-validation (every 60s) checks
  is_active/role from DB, invalidates sessions of deactivated users and
  refreshes role changes without requiring re-login
- PolicyRepository: add optional tenantId to findQuorumPolicy/findVotePolicy
  for defense-in-depth tenant isolation on policy lookups
- MotionFinderTrait: add tenantId to findOpenForProjector,
  findLastClosedForProjector, findNextNotOpened
- MeetingRepository: add tenantId to findTitle
- AttendanceRepository: fix ::int truncation on present_weight (now ::float8)
  — fractional voting_power was silently truncated in dashboard summaries
- Update all callers across 10 controllers/services to pass tenantId

All 2739 tests green.

+++ app/Repository/AttendanceRepository.php
@@ -103,7 +103,7 @@ class AttendanceRepository extends AbstractRepository {
         $row = $this->selectOne(
             "SELECT
                 COUNT(*) FILTER (WHERE a.mode IN ('present','remote'))::int AS present_count,
-                COALESCE(SUM(COALESCE(m.voting_power, 1.0)) FILTER (WHERE a.mode IN ('present','remote')),0)::int AS present_weight
+                COALESCE(SUM(COALESCE(m.voting_power, 1.0)) FILTER (WHERE a.mode IN ('present','remote')),0)::float8 AS present_weight
              FROM attendances a
              JOIN members m ON m.id = a.member_id
              WHERE a.tenant_id = :tid AND a.meeting_id = :mid",
